modularized volumes from nodes into NodeVolumes
